Change stragegy for determining when pubsub is empty
Counter to expectation, Google pubsub pull does not always return the
requested max number of elements if the pubsub queue has a greater
number of elements than the number requested.

The new strategy for determining if the queue is empty is if we receive
two successive pulls that have less than the requested amount.