Stop sending errors to airbrake if too many or rate limited
There are two scaling issues in the airbrake notifier

1. Airbrake is rate limiting our API requests
2. We have too many errors to send to airbrake, whether or not they are
   rate limiting us

In the first case, it does not make sense to send further requests to
airbrake in the current Lambda execution since they will likely be
rejected. This commit stops the requests if we hit the rate limit.

In the second case, errors are likely repeats. We do not get additional
information by sending more to airbrake. The ones represented in their
systems will be enough to work on. We will likely see the othe errors
once we have cleared the current ones. This commit limits the lambda to
sending 50 notifications per lambda execution.